move plugins to their own directory
