+++ example/lib/plugins/plugin_scalebar.dart
